pyproject.toml
[tool.poetry]
name = "alloff_grpc_protos"
version = "0.3.2"
description = ""
authors = ["Seokmin Lee <essemfly@gmail.com>"]
packages = [
    { include = "gen" },
]

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.45.0"
grpcio-tools = "^1.45.0"
2to3 = "^1.0"
djangogrpcframework = "^0.2.1"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
